+++ test/dart_dice_parser_test.dart
@@ -11,6 +11,7 @@ void main() {
   DiceParser staticDiceParser;
   // dice parser whose random returns incrementing ints from 0, 1, 2... (rolls 1,2,3...)
   DiceParser incrementingDiceParser;
+  DiceParser explodeDiceParser;
 
   setUp(() async {
     var staticMockRandom = MockRandom();
@@ -28,6 +29,29 @@ void main() {
     staticDiceParser = DiceParser(diceRoller: DiceRoller(staticMockRandom));
     incrementingDiceParser =
         DiceParser(diceRoller: DiceRoller(incrementingMockRandom));
+
+    var explodeMockRandom = MockRandom();
+
+    var explodeResponses = [
+      // 1st roll: 2,3,4,4 (explodes 2)
+      1,
+      2,
+      3,
+      3,
+      // 2nd roll: 4,4 (explodes 2)
+      3,
+      3,
+      // 3rd roll: 4,2 (explodes 1)
+      3,
+      1,
+      // final roll: 2
+      1,
+      0, 0, 0, 0, 0, 0, 0, 0
+    ];
+    when(explodeMockRandom.nextInt(argThat(inInclusiveRange(1, 1000))))
+        .thenAnswer((_) => explodeResponses.removeAt(0));
+
+    explodeDiceParser = DiceParser(diceRoller: DiceRoller(explodeMockRandom));
   });
 
   group("arithmetic", () {
@@ -57,7 +81,40 @@ void main() {
     });
   });
 
-  group("drop high/low:", () {
+  group("roll operations", () {
+    test("count >", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6#>3"), equals(1));
+    });
+    test("count <", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6#<4"), equals(3));
+    });
+    test("count =", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6#=2"), equals(1));
+    });
+    test("count =", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6#>5"), equals(0));
+    });
+
+    test("count #", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6#"), equals(4));
+    });
+    test("count # after drop", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6-<2#"), equals(3));
+    });
+  });
+  group("roll modifiers", () {
     test("drop high", () {
       // mocked responses should return rolls of 1,2,3,4
       //
@@ -97,6 +154,31 @@ void main() {
     });
     test("can drop more than rolled", () {
       expect(incrementingDiceParser.roll("3d6-H4"), equals(0));
+    });
+    test("drop >3", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6->3"), equals(6));
+    });
+    test("drop <3", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6-<3"), equals(7));
+    });
+    test("drop =3", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6-=3"), equals(7));
+    });
+    test("clamp >3", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6C>3"), equals(9));
+    });
+    test("clamp <3", () {
+      // mocked responses should return rolls of 1,2,3,4
+      //
+      expect(incrementingDiceParser.roll("4d6C<3"), equals(13));
     });
   });
 
@@ -163,6 +245,28 @@ void main() {
     test("fudge", () {
       var input = "4dF";
       expect(staticDiceParser.roll(input), equals(-4));
+    });
+    test("exploding dice", () {
+      // 1st roll: 2,3,4,4 (explodes 2)
+      // 2nd roll: 4,4 (explodes 2)
+      // 3rd roll: 4,2 (explodes 1)
+      // final roll: 2
+      expect(explodeDiceParser.roll("4d!4"), equals(29));
+    });
+
+    test("exploding dice and count", () {
+      // explode, then count 4's
+      expect(explodeDiceParser.roll("4d!4#=4"), equals(5));
+    });
+    test("exploding dice and count variation", () {
+      // explode, then drop less-than-4, then count total
+      expect(explodeDiceParser.roll("4d!4-<4#"), equals(5));
+    });
+
+    test("exploding dice", () {
+      // 1st roll: 2,3,4,4 (explodes 2)
+      // 2nd roll: 4,4 (explodes 2) (should stop here)
+      expect(explodeDiceParser.roll("4d!!4"), equals(21));
     });
 
     test("invalid dice str", () {